requirements.txt
pandas
google-cloud-bigquery
google-generativeai>=0.8.0
db-dtypes
yfinance
cloudscraper
html5lib
lxml
duckduckgo-search
streamlit>=1.35.0
boto3
